requirements.txt

protobuf>=3.6
grpcio-tools

pytest>=4.3
pytest-asyncio>=0.9
